requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
beautifulsoup4==4.4.1
blinker==1.3
Brlapi==0.6.4
cffi==1.12.3
chardet==2.3.0
checkbox-support==0.22
click==6.2
colorama==0.3.7
command-not-found==0.3
cryptography==2.7
defer==1.0.6
feedparser==5.1.3
fido2==0.7.0
guacamole==0.9.2
html5lib==0.999
httplib2==0.9.1
idna==2.0
Jinja2==2.8
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
MarkupSafe==0.23
mpmath==1.1.0
oauthlib==1.0.3
onboard==1.2.0
padme==1.1.1
pexpect==4.0.1
Pillow==3.1.2
plainbox==0.25
ptyprocess==0.5
pyasn1==0.1.9
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.8.2
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.3.0
pyOpenSSL==0.15.1
pyparsing==2.0.3
pyscard==1.9.2
pyserial==3.4
python-apt==1.1.0b1+ubuntu0.16.4.2
python-debian==0.1.27
python-systemd==231
pyusb==1.0.0b2
pyxdg==0.25
pyzmq==18.1.0
reportlab==3.3.0
requests==2.9.1
sessioninstaller==0.0.0
six==1.12.0
SQLAlchemy==1.0.11
-e git+https://github.com/sympy/sympy.git@40da07b3a8ca3b315aaf25ec4ac21dfe4364a097#egg=sympy
system-service==0.3
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.13.1
usb-creator==0.3.0
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
yubikey-manager==3.0.0
